unify GAP8 tilers (excluding board_L2)
